+++ CHANGELOG.md
@@ -99,7 +99,7 @@
 
 - Add back the dependency on build-essential cookbook to restore Chef Infra Client < 14 compatibility
 - Remove the complex compile_time logic to install the gem in the default recipe as compile_time is available in Chef Infra Client 12.1+
-- Added 3 new resources for Consule 1.5+
+- Added 3 new resources for Consul 1.5+
   - consul_token
   - consul_policy
   - consul_role